Rename strategies.py to hypothesis_strategies.py
